Refactor component API
This is a huge change, and it should be made until it is too
late. There are several reasons for this change:

- Use the same naming scheme as web api (not so important)
- Different semantics for `mount`, it is extremely important that
attached methods are invoked in top to bottom order, and detached is in
bottom to up order. We don't have problems with refs, because refs are
assigned on instantiation, and attached methods are invoked when
everything is instantiated. We shouldn't copy broken behavior of other
libraries.

This patch also slightly changes context API. Context values modified by
the component are available only to descendants, component can only
access context values that came from parent nodes. Context modifications
also won't trigger update to the component that made context changes,
because it can't use this values.