Change context API, optimize for read perf
Context API is now slightly simpler, and instead of optimizing for write
performance it is now optimized for read performance. It seems that in
real applications read operations is way much more common, so we just
merge objects when context is updated, and read operations doesn't
need to traverse context tree to find all values.