Change children normalization to be more restrictive
It doesn't support recursive trees anymore, only 2-depth trees. This
change doesn't have any effect on performance, it is changed to remove
all underlying magic for implicit positional keys, now it is so much
easier to understand how positional keys are assigned when children
are normalized.

Also, now it is required to have explicit keys on all nodes for
nested arrays (similar to React), so there is no need to normalize
nested arrays.

There are probably will be extremely rare situations when it will be
required to write slightly more complicated code because of this
change, but this change doesn't make it impossible, just slightly more
complicated. Most of the other vdom libraries are using way much
simpler normalization algorithms and nobody even noticing that they
don't support such things. React does way much better job, but at the
same time it forces to use explicit keys in all nodes in nested
arrays, so there is no point to overcomplicate algorithm and then
print warnings in the console that this feature should be avoided.