Remove VNode ref API
React ref API is broken by design. I think that it is better to have
correct behavior for majority of use cases and make it slightly harder
to deal with some extremely rare use cases that require finding
Component/DOM instances (it is actually quite easy to find instances
from VNodes in ivi). The other solution would be to add more lifecycle
hooks, but it will significantly overcomplicate API, so it is better
just to remove ref API.