Major improvements of core algorithms and data structures
I guess that is how next generation of efficient vdom libraries will
look like :) There are still some optimizations that needs to be done,
new algo just passes all tests and it is already beating kivi dbmon
performance with way much simpler dbmon implementation that doesn't
use any tricks.

This change also makes it that simple and idiomatic code with map
functions is more efficient than preallocating arrays and filling them
with for-loops.

Also, it handles cases with nested/null children without any perf
costs, something that benchmarks are usually ignoring, but it is
everywhere in real apps.

Code size is also reduced, it is just so much better in every way :)