Immediately execute useEffect() hooks
`useEffect()` hook is now immediately executed when it is invoked in the
 internal component "update" function. It makes it fully deterministic
to prevent unexpected bugs when `useEffect()` is used to listen some
observable value that can generate an event before microtask with an
effect is executed. When it is immediately executed, it will guarantee
that we don't miss any events generated by an observable value.