Optimize stateful tree
Switch from a single object shape for stateful nodes to two shapes. It
will make some call-sites polymorphic, but it is not a problem, because
it will have a degree of polymorphism 2.

Two object shapes would allow us to remove creating addional object to
store component state and removes one level of indirection when
accessing it. In chromium-based engines it is going to reduce
memory usagy by 16 bytes per component node.