Move `@ivi/template-compiler` into `ivi` package
Since ivi already has a hard dependency on this package, it is better to
just move it inside and remove all peerDependencies from different
template parsers.

Also, it solves an issue that apps need to add template compiler to
their dependencies, because template parsers are using it as a peerDep.